Create a function that recives parent and child
Pass the parent to append the child

Create a class with method that is responsible for conditioning
appending element.

Append an element if there is an object or two and
if the input length is greater than or equal to 4 to strictly prohibit
input value lengths that are less than 4.

Use imported function to namespace.
Use project arr to to get the value of project name and project index
Use these values to put it into another function to successfully append
them containing the project index and project name