Set 1 project as a key and array as it's value
In the local storage i have seen that it's possible to have a single
array and put multiple objects in it. i just thought well other's did it
and so i have come to try it myself.

Instead of setting a Project index every time i create them. it's better
this way. only bit of changes like removing the for loop for the getting
from localStorage and parsing them.